+++ R/mergeUniformCellsClusters.R
@@ -21,6 +21,13 @@
 #'   known for not being uniform. Useful to restart the *merging* process after
 #'   an interruption.
 #' @param cores number of cores to use. Default is 1.
+#' @param optimizeForSpeed Boolean; when `TRUE` `COTAN` tries to use the `torch`
+#'   library to run the matrix calculations. Otherwise, or when the library is
+#'   not available will run the slower legacy code
+#' @param deviceStr On the `torch` library enforces which device to use to run
+#'   the calculations. Possible values are `"cpu"` to us the system *CPU*,
+#'   `"cuda"` to use the system *GPUs* or something like `"cuda:0"` to restrict
+#'   to a specific device
 #' @param useDEA Boolean indicating whether to use the *DEA* to define the
 #'   distance; alternatively it will use the average *Zero-One* counts, that is
 #'   faster but less precise.
@@ -84,18 +91,19 @@
 #' ##
 #'
 #' splitList <- cellsUniformClustering(objCOTAN, cores = 6L,
+#'                                     optimizeForSpeed = TRUE,
+#'                                     deviceStr = "cuda",
 #'                                     initialResolution = 0.8,
 #'                                     GDIThreshold = 1.46, saveObj = FALSE)
 #'
 #' clusters <- splitList[["clusters"]]
 #'
 #' firstCluster <- getCells(objCOTAN)[clusters %in% clusters[[1L]]]
-#' checkClusterUniformity(objCOTAN,
-#'                        GDIThreshold = 1.46,
-#'                        cluster = clusters[[1L]],
-#'                        cells = firstCluster,
-#'                        cores = 6L,
-#'                        saveObj = FALSE)
+#' firstClusterIsUniform <-
+#'   checkClusterUniformity(objCOTAN, GDIThreshold = 1.46,
+#'                          cluster = clusters[[1L]], cells = firstCluster,
+#'                          cores = 6L, optimizeForSpeed = TRUE,
+#'                          deviceStr = "cuda", saveObj = FALSE)[["isUniform"]]
 #'
 #' objCOTAN <- addClusterization(objCOTAN,
 #'                               clName = "split",
@@ -112,6 +120,8 @@
 #'                                         batchSize = 5L,
 #'                                         clusters = clusters,
 #'                                         cores = 6L,
+#'                                         optimizeForSpeed = TRUE,
+#'                                         deviceStr = "cpu",
 #'                                         distance = "cosine",
 #'                                         hclustMethod = "ward.D2",
 #'                                         saveObj = FALSE)
@@ -132,6 +142,8 @@ mergeUniformCellsClusters <- function(objCOTAN,
                                       batchSize = 10L,
                                       notMergeable = NULL,
                                       cores = 1L,
+                                      optimizeForSpeed = TRUE,
+                                      deviceStr = "cuda",
                                       useDEA = TRUE,
                                       distance = NULL,
                                       hclustMethod = "ward.D2",
@@ -173,9 +185,8 @@ mergeUniformCellsClusters <- function(objCOTAN,
 
   iter <- 0L
   allCheckResults <- data.frame()
-  errorCheckResults <-
-    list("isUniform" = FALSE, "fractionAbove" = NA, "firstPercentile" = NA)
-
+  errorCheckResults <- list("isUniform" = FALSE, "fractionAbove" = NA,
+                            "firstPercentile" = NA, "size" = NA)
 
   testPairListMerge <- function(pList) {
     logThis(paste0("Clusters pairs for merging:\n",
@@ -208,16 +219,19 @@ mergeUniformCellsClusters <- function(objCOTAN,
 
       checkResults <- tryCatch(
         checkClusterUniformity(objCOTAN,
-                               cluster = mergedClName,
+                               clusterName = mergedClName,
                                cells = mergedCluster,
                                GDIThreshold = GDIThreshold,
                                cores = cores,
+                               optimizeForSpeed = optimizeForSpeed,
+                               deviceStr = deviceStr,
                                saveObj = saveObj,
                                outDir = mergeOutDir),
         error = function(err) {
           logThis(paste("While checking cluster uniformity", err),
                   logLevel = 0L)
           logThis("Marking pair as not mergable", logLevel = 1L)
+          errorCheckResults[["size"]] <- length(mergedCluster)
           return(errorCheckResults)
         })
 
@@ -250,8 +264,7 @@ mergeUniformCellsClusters <- function(objCOTAN,
     oldNumClusters <- length(unique(outputClusters))
 
     clDist <- distancesBetweenClusters(objCOTAN, clusters = outputClusters,
-                                       useDEA = useDEA, cores = cores,
-                                       distance = distance)
+                                       useDEA = useDEA, distance = distance)
     gc()
 
     if (isTRUE(saveObj)) tryCatch({
@@ -354,7 +367,7 @@ mergeUniformCellsClusters <- function(objCOTAN,
   }
 
   outputCoexDF <-
-    tryCatch(DEAOnClusters(objCOTAN, clusters = outputClusters, cores = cores),
+    tryCatch(DEAOnClusters(objCOTAN, clusters = outputClusters),
              error = function(err) {
                logThis(paste("Calling DEAOnClusters", err), logLevel = 0L)
                return(NULL)
@@ -363,7 +376,7 @@ mergeUniformCellsClusters <- function(objCOTAN,
   c(outputClusters, outputCoexDF) %<-% tryCatch(
     reorderClusterization(objCOTAN, clusters = outputClusters,
                           coexDF = outputCoexDF, reverse = FALSE,
-                          keepMinusOne = FALSE, useDEA = useDEA, cores = cores,
+                          keepMinusOne = FALSE, useDEA = useDEA,
                           distance = distance, hclustMethod = hclustMethod),
     error = function(err) {
       logThis(paste("Calling reorderClusterization", err), logLevel = 0L)