+++ tests/testthat/test-clean-method.R
@@ -25,7 +25,7 @@ test_that("Clean on test dataset", {
 
   obj <- clean(obj)
 
-  obj <- estimateDispersionBisection(obj, cores = 12L)
+  obj <- estimateDispersionBisection(obj, cores = 6L)
 
   raw.norm <- readRDS(file.path(getwd(), "raw.norm.test.RDS"))
   lambda <- readRDS(file.path(getwd(), "lambda.test.RDS"))
@@ -33,7 +33,7 @@ test_that("Clean on test dataset", {
   nu <- readRDS(file.path(getwd(), "nu.test.RDS"))
 
   genes.names.test <- readRDS(file.path(getwd(), "genes.names.test.RDS"))
-  cells.names.test <- readRDS(file.path(getwd(), "cell.names.test.RDS"))
+  cells.names.test <- readRDS(file.path(getwd(), "cells.names.test.RDS"))
 
   expect_equal(getNormalizedData(obj)[genes.names.test, cells.names.test],
                raw.norm, tolerance = 1.0e-14, ignore_attr = FALSE)